+++ src/main/java/com/tech/padawan/financialmanager/goal/dto/UpdateSaveAmountDTO.java
@@ -3,10 +3,12 @@ package com.tech.padawan.financialmanager.goal.dto;
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigDecimal;
+
 public record UpdateSaveAmountDTO(
         @NotNull(message = "Value is required")
         @Min(value = 0, message = "Value must be positive")
-        Double value
+        BigDecimal value
 ) {
 
 }
